media_processor/pyproject.toml
[tool.pytest.ini_options]
pythonpath = [
    ".", "src",
]
asyncio_mode="auto"

[tool.poetry]
name = "test-ruff"
version = "1"
description = "fuck you"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ruff = "^0.5.7"
mypy = "^1.11.1"
black = "^24.8.0"
fastapi = "^0.112.0"
asyncpg = "^0.29.0"
asyncio = "^3.4.3"
uvicorn = "^0.30.5"
aiokafka = "^0.11.0"
aiobotocore = "^2.13.2"
python-dotenv = "^1.0.1"
types-aiobotocore = "^2.13.2"
pydantic-settings = "^2.4.0"
python-multipart = "^0.0.9"
minio = "^7.2.8"
python-ffmpeg = "^2.0.12"
streaming-form-data = "^1.16.0"
aiofiles = "^24.1.0"
py-spy = "^0.3.14"
asyncffmpeg = "^1.2.0"
asynccpu = "^1.2.2"


[tool.poetry.group.dev.dependencies]
pyinstrument = "^4.7.2"
pytest = "^8.3.2"
pytest-aiohttp = "^1.0.5"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
